Struct aws_sdk_waf::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AWSWAF_20150824.

This client allows ergonomic access to a AWSWAF_20150824-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn create_byte_match_set(&self) -> CreateByteMatchSet<C, M, R>

pub fn create_geo_match_set(&self) -> CreateGeoMatchSet<C, M, R>

pub fn create_ip_set(&self) -> CreateIPSet<C, M, R>

pub fn create_rate_based_rule(&self) -> CreateRateBasedRule<C, M, R>

pub fn create_regex_match_set(&self) -> CreateRegexMatchSet<C, M, R>

pub fn create_regex_pattern_set(&self) -> CreateRegexPatternSet<C, M, R>

pub fn create_rule(&self) -> CreateRule<C, M, R>

pub fn create_rule_group(&self) -> CreateRuleGroup<C, M, R>

pub fn create_size_constraint_set(&self) -> CreateSizeConstraintSet<C, M, R>

pub fn create_sql_injection_match_set(
&self
) -> CreateSqlInjectionMatchSet<C, M, R>

pub fn create_web_acl(&self) -> CreateWebACL<C, M, R>

pub fn create_web_acl_migration_stack(
&self
) -> CreateWebACLMigrationStack<C, M, R>

pub fn create_xss_match_set(&self) -> CreateXssMatchSet<C, M, R>

pub fn delete_byte_match_set(&self) -> DeleteByteMatchSet<C, M, R>

pub fn delete_geo_match_set(&self) -> DeleteGeoMatchSet<C, M, R>

pub fn delete_ip_set(&self) -> DeleteIPSet<C, M, R>

pub fn delete_logging_configuration(
&self
) -> DeleteLoggingConfiguration<C, M, R>

pub fn delete_permission_policy(&self) -> DeletePermissionPolicy<C, M, R>

pub fn delete_rate_based_rule(&self) -> DeleteRateBasedRule<C, M, R>

pub fn delete_regex_match_set(&self) -> DeleteRegexMatchSet<C, M, R>

pub fn delete_regex_pattern_set(&self) -> DeleteRegexPatternSet<C, M, R>

pub fn delete_rule(&self) -> DeleteRule<C, M, R>

pub fn delete_rule_group(&self) -> DeleteRuleGroup<C, M, R>

pub fn delete_size_constraint_set(&self) -> DeleteSizeConstraintSet<C, M, R>

pub fn delete_sql_injection_match_set(
&self
) -> DeleteSqlInjectionMatchSet<C, M, R>

pub fn delete_web_acl(&self) -> DeleteWebACL<C, M, R>

pub fn delete_xss_match_set(&self) -> DeleteXssMatchSet<C, M, R>

pub fn get_byte_match_set(&self) -> GetByteMatchSet<C, M, R>

pub fn get_change_token(&self) -> GetChangeToken<C, M, R>

pub fn get_change_token_status(&self) -> GetChangeTokenStatus<C, M, R>

pub fn get_geo_match_set(&self) -> GetGeoMatchSet<C, M, R>

pub fn get_ip_set(&self) -> GetIPSet<C, M, R>

pub fn get_logging_configuration(&self) -> GetLoggingConfiguration<C, M, R>

pub fn get_permission_policy(&self) -> GetPermissionPolicy<C, M, R>

pub fn get_rate_based_rule(&self) -> GetRateBasedRule<C, M, R>

pub fn get_rate_based_rule_managed_keys(
&self
) -> GetRateBasedRuleManagedKeys<C, M, R>

pub fn get_regex_match_set(&self) -> GetRegexMatchSet<C, M, R>

pub fn get_regex_pattern_set(&self) -> GetRegexPatternSet<C, M, R>

pub fn get_rule(&self) -> GetRule<C, M, R>

pub fn get_rule_group(&self) -> GetRuleGroup<C, M, R>

pub fn get_sampled_requests(&self) -> GetSampledRequests<C, M, R>

pub fn get_size_constraint_set(&self) -> GetSizeConstraintSet<C, M, R>

pub fn get_sql_injection_match_set(&self) -> GetSqlInjectionMatchSet<C, M, R>

pub fn get_web_acl(&self) -> GetWebACL<C, M, R>

pub fn get_xss_match_set(&self) -> GetXssMatchSet<C, M, R>

pub fn list_activated_rules_in_rule_group(
&self
) -> ListActivatedRulesInRuleGroup<C, M, R>

pub fn list_byte_match_sets(&self) -> ListByteMatchSets<C, M, R>

pub fn list_geo_match_sets(&self) -> ListGeoMatchSets<C, M, R>

pub fn list_ip_sets(&self) -> ListIPSets<C, M, R>

pub fn list_logging_configurations(&self) -> ListLoggingConfigurations<C, M, R>

pub fn list_rate_based_rules(&self) -> ListRateBasedRules<C, M, R>

pub fn list_regex_match_sets(&self) -> ListRegexMatchSets<C, M, R>

pub fn list_regex_pattern_sets(&self) -> ListRegexPatternSets<C, M, R>

pub fn list_rule_groups(&self) -> ListRuleGroups<C, M, R>

pub fn list_rules(&self) -> ListRules<C, M, R>

pub fn list_size_constraint_sets(&self) -> ListSizeConstraintSets<C, M, R>

pub fn list_sql_injection_match_sets(
&self
) -> ListSqlInjectionMatchSets<C, M, R>

pub fn list_subscribed_rule_groups(&self) -> ListSubscribedRuleGroups<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn list_web_ac_ls(&self) -> ListWebACLs<C, M, R>

pub fn list_xss_match_sets(&self) -> ListXssMatchSets<C, M, R>

pub fn put_logging_configuration(&self) -> PutLoggingConfiguration<C, M, R>

pub fn put_permission_policy(&self) -> PutPermissionPolicy<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_byte_match_set(&self) -> UpdateByteMatchSet<C, M, R>

pub fn update_geo_match_set(&self) -> UpdateGeoMatchSet<C, M, R>

pub fn update_ip_set(&self) -> UpdateIPSet<C, M, R>

pub fn update_rate_based_rule(&self) -> UpdateRateBasedRule<C, M, R>

pub fn update_regex_match_set(&self) -> UpdateRegexMatchSet<C, M, R>

pub fn update_regex_pattern_set(&self) -> UpdateRegexPatternSet<C, M, R>

pub fn update_rule(&self) -> UpdateRule<C, M, R>

pub fn update_rule_group(&self) -> UpdateRuleGroup<C, M, R>

pub fn update_size_constraint_set(&self) -> UpdateSizeConstraintSet<C, M, R>

pub fn update_sql_injection_match_set(
&self
) -> UpdateSqlInjectionMatchSet<C, M, R>

pub fn update_web_acl(&self) -> UpdateWebACL<C, M, R>

pub fn update_xss_match_set(&self) -> UpdateXssMatchSet<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self