Module aws_sdk_waf::error

Modules

waf_bad_request_exception

See WafBadRequestException

waf_disallowed_name_exception

See WafDisallowedNameException

waf_entity_migration_exception

See WafEntityMigrationException

waf_internal_error_exception

See WafInternalErrorException

waf_invalid_account_exception

See WafInvalidAccountException

waf_invalid_operation_exception

See WafInvalidOperationException

waf_invalid_parameter_exception

See WafInvalidParameterException

waf_invalid_permission_policy_exception

See WafInvalidPermissionPolicyException

waf_invalid_regex_pattern_exception

See WafInvalidRegexPatternException

waf_limits_exceeded_exception

See WafLimitsExceededException

waf_non_empty_entity_exception

See WafNonEmptyEntityException

waf_nonexistent_container_exception

See WafNonexistentContainerException

waf_nonexistent_item_exception

See WafNonexistentItemException

waf_referenced_item_exception

See WafReferencedItemException

waf_service_linked_role_error_exception

See WafServiceLinkedRoleErrorException

waf_stale_data_exception

See WafStaleDataException

waf_subscription_not_found_exception

See WafSubscriptionNotFoundException

waf_tag_operation_exception

See WafTagOperationException

waf_tag_operation_internal_error_exception

See WafTagOperationInternalErrorException

Structs

CreateByteMatchSetError

CreateGeoMatchSetError

CreateIPSetError

CreateRateBasedRuleError

CreateRegexMatchSetError

CreateRegexPatternSetError

CreateRuleError

CreateRuleGroupError

CreateSizeConstraintSetError

CreateSqlInjectionMatchSetError

CreateWebACLError

CreateWebACLMigrationStackError

CreateXssMatchSetError

DeleteByteMatchSetError

DeleteGeoMatchSetError

DeleteIPSetError

DeleteLoggingConfigurationError

DeletePermissionPolicyError

DeleteRateBasedRuleError

DeleteRegexMatchSetError

DeleteRegexPatternSetError

DeleteRuleError

DeleteRuleGroupError

DeleteSizeConstraintSetError

DeleteSqlInjectionMatchSetError

DeleteWebACLError

DeleteXssMatchSetError

GetByteMatchSetError

GetChangeTokenError

GetChangeTokenStatusError

GetGeoMatchSetError

GetIPSetError

GetLoggingConfigurationError

GetPermissionPolicyError

GetRateBasedRuleError

GetRateBasedRuleManagedKeysError

GetRegexMatchSetError

GetRegexPatternSetError

GetRuleError

GetRuleGroupError

GetSampledRequestsError

GetSizeConstraintSetError

GetSqlInjectionMatchSetError

GetWebACLError

GetXssMatchSetError

ListActivatedRulesInRuleGroupError

ListByteMatchSetsError

ListGeoMatchSetsError

ListIPSetsError

ListLoggingConfigurationsError

ListRateBasedRulesError

ListRegexMatchSetsError

ListRegexPatternSetsError

ListRuleGroupsError

ListRulesError

ListSizeConstraintSetsError

ListSqlInjectionMatchSetsError

ListSubscribedRuleGroupsError

ListTagsForResourceError

ListWebACLsError

ListXssMatchSetsError

PutLoggingConfigurationError

PutPermissionPolicyError

TagResourceError

UntagResourceError

UpdateByteMatchSetError

UpdateGeoMatchSetError

UpdateIPSetError

UpdateRateBasedRuleError

UpdateRegexMatchSetError

UpdateRegexPatternSetError

UpdateRuleError

UpdateRuleGroupError

UpdateSizeConstraintSetError

UpdateSqlInjectionMatchSetError

UpdateWebACLError

UpdateXssMatchSetError

WafBadRequestException

WafDisallowedNameException

The name specified is invalid.

WafEntityMigrationException

The operation failed due to a problem with the migration. The failure cause is provided in the exception, in the MigrationErrorType:

WafInternalErrorException

The operation failed because of a system problem, even though the request was valid. Retry your request.

WafInvalidAccountException

The operation failed because you tried to create, update, or delete an object by using an invalid account identifier.

WafInvalidOperationException

The operation failed because there was nothing to do. For example:

WafInvalidParameterException

The operation failed because AWS WAF didn't recognize a parameter in the request. For example:

WafInvalidPermissionPolicyException

The operation failed because the specified policy is not in the proper format.

WafInvalidRegexPatternException

The regular expression (regex) you specified in RegexPatternString is invalid.

WafLimitsExceededException

The operation exceeds a resource limit, for example, the maximum number of WebACL objects that you can create

WafNonEmptyEntityException

The operation failed because you tried to delete an object that isn't empty. For example:

WafNonexistentContainerException

The operation failed because you tried to add an object to or delete an object from another object that doesn't exist. For example:

WafNonexistentItemException

The operation failed because the referenced object doesn't exist.

WafReferencedItemException

The operation failed because you tried to delete an object that is still in use. For example:

WafServiceLinkedRoleErrorException

AWS WAF is not able to access the service linked role. This can be caused by a previous PutLoggingConfiguration request, which can lock the service linked role for about 20 seconds. Please try your request again. The service linked role can also be locked by a previous DeleteServiceLinkedRole request, which can lock the role for 15 minutes or more. If you recently made a DeleteServiceLinkedRole, wait at least 15 minutes and try the request again. If you receive this same exception again, you will have to wait additional time until the role is unlocked.

WafStaleDataException

The operation failed because you tried to create, update, or delete an object by using a change token that has already been used.

WafSubscriptionNotFoundException

The specified subscription does not exist.

WafTagOperationException

WafTagOperationInternalErrorException

Enums

CreateByteMatchSetErrorKind

CreateGeoMatchSetErrorKind

CreateIPSetErrorKind

CreateRateBasedRuleErrorKind

CreateRegexMatchSetErrorKind

CreateRegexPatternSetErrorKind

CreateRuleErrorKind

CreateRuleGroupErrorKind

CreateSizeConstraintSetErrorKind

CreateSqlInjectionMatchSetErrorKind

CreateWebACLErrorKind

CreateWebACLMigrationStackErrorKind

CreateXssMatchSetErrorKind

DeleteByteMatchSetErrorKind

DeleteGeoMatchSetErrorKind

DeleteIPSetErrorKind

DeleteLoggingConfigurationErrorKind

DeletePermissionPolicyErrorKind

DeleteRateBasedRuleErrorKind

DeleteRegexMatchSetErrorKind

DeleteRegexPatternSetErrorKind

DeleteRuleErrorKind

DeleteRuleGroupErrorKind

DeleteSizeConstraintSetErrorKind

DeleteSqlInjectionMatchSetErrorKind

DeleteWebACLErrorKind

DeleteXssMatchSetErrorKind

GetByteMatchSetErrorKind

GetChangeTokenErrorKind

GetChangeTokenStatusErrorKind

GetGeoMatchSetErrorKind

GetIPSetErrorKind

GetLoggingConfigurationErrorKind

GetPermissionPolicyErrorKind

GetRateBasedRuleErrorKind

GetRateBasedRuleManagedKeysErrorKind

GetRegexMatchSetErrorKind

GetRegexPatternSetErrorKind

GetRuleErrorKind

GetRuleGroupErrorKind

GetSampledRequestsErrorKind

GetSizeConstraintSetErrorKind

GetSqlInjectionMatchSetErrorKind

GetWebACLErrorKind

GetXssMatchSetErrorKind

ListActivatedRulesInRuleGroupErrorKind

ListByteMatchSetsErrorKind

ListGeoMatchSetsErrorKind

ListIPSetsErrorKind

ListLoggingConfigurationsErrorKind

ListRateBasedRulesErrorKind

ListRegexMatchSetsErrorKind

ListRegexPatternSetsErrorKind

ListRuleGroupsErrorKind

ListRulesErrorKind

ListSizeConstraintSetsErrorKind

ListSqlInjectionMatchSetsErrorKind

ListSubscribedRuleGroupsErrorKind

ListTagsForResourceErrorKind

ListWebACLsErrorKind

ListXssMatchSetsErrorKind

PutLoggingConfigurationErrorKind

PutPermissionPolicyErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateByteMatchSetErrorKind

UpdateGeoMatchSetErrorKind

UpdateIPSetErrorKind

UpdateRateBasedRuleErrorKind

UpdateRegexMatchSetErrorKind

UpdateRegexPatternSetErrorKind

UpdateRuleErrorKind

UpdateRuleGroupErrorKind

UpdateSizeConstraintSetErrorKind

UpdateSqlInjectionMatchSetErrorKind

UpdateWebACLErrorKind

UpdateXssMatchSetErrorKind